mail: add Header.MessageID and Header.MsgIDList
Those are defined in RFC 5322 section 3.6.4. We need to provide helpers
because the fields may contain CFWS.

We use headerParser introduced in the previous commit.

Message identifiers are returned without the angle brackets, because the
RFC says:

> Semantically, the angle bracket characters are not part of the
> msg-id; the msg-id is what is contained between the two angle bracket
> characters.

Also, it's easier for the caller to add back the angle brackets if
needed rather than stripping them.

MsgIDList isn't the best name ever. The RFC defines the message-id token
for the Message-ID header field and the msg-id token for message
identifiers, hence this name. I wonder if we can come up with something
better.

Maybe we should add Header.SetMessageID and Header.SetMsgIDList. Maybe
we should remove our existing GenerateMessageID function, and add it to
Header.

[1]: https://tools.ietf.org/html/rfc5322#section-3.6.4

References: https://github.com/emersion/go-message/issues/49