Remove hz-gb-2312 encoding as it can crash
Upstream go's x/text/encoding package has a bug that makes the
'hz-gb-2312' encoding crash (https://github.com/golang/go/issues/35118)

This was reported to go-message in
https://github.com/emersion/go-message/issues/95

This commit removes the offending encoding so that go-message can't crash if
people import the charset package. It should be reverted once the upstream
package is fixed by the go developers.