message: introduce Entity.Walk
Entity.Walk is similar to filepath.Walk [1]: it iterates over all parts
in the entity tree.

There is no equivalent to SkipDir, because the whole tree will need to
be read anyway (ie. the performance benefits of filepath.SkipDir cannot
apply here).

WalkFunc paths are different from IMAP part paths: the root part
doesn't have a path equal to {1}, and indices start at 0. The former is
necessary to disambiguate between the root part and its first child if
it's multipart.

[1]: https://golang.org/pkg/path/filepath/#Walk

Closes: https://github.com/emersion/go-message/issues/88