Export Unknown{Charset,Encoding}Errors
This allow us to use the go 1.13 error handling facility which is a bit cleaner
as it allows for a nested error lookup with

```
err = someTwoCallDeepGoMessageFunc()
        if errors.As(err, new(message.UnknownEncodingError)){
                handleIt()
        }
```

Else it gets really clunky, as the caller has to unwrap the error to the exact error
returned by go-message and then using message.IsUnknownEncoding(err).

To keep backwards compatibility I've left the helpers in, so there should be no
breaking changes.