charset: fix htmlindex fallback
In some cases, ianaindex would return a nil Encoding to mean that it
doesn't support the charset *and* a nil error to mean that the supplied
charset is a valid IANA charset. This isn't great, and [1] contains an
idea to fix this bad API.

In the meantime, check enc == nil instead err != nil.

[1]: https://go-review.googlesource.com/c/text/+/212077/5#message-59f8d7185d7f10e825a79541046b0aff668d7494

Fixes: d053f95466fb ("charset: remove charsets supported by htmlindex")
Closes: https://github.com/emersion/go-message/issues/104