operate on a copy of the header for write methods
If callers pass the same header to CreateSingleInlineWriter or similar methods,
the header gets modified in place, which means that the header of the caller
and the header of go-message go out of sync and subsequent operations on the
same header fail.
Operate on a copy in all paths that modify the header to avoid this issue.