Move header size restriction from textproto to message
It's not really useful to individually check the line length or the
number of fields, what we really care about is the total header size.
This is also what net/http checks for.

Remove size checks from textproto, since callers can implement them. Use
a variant of io.LimitedReader in message instead.